+++ src/test/kotlin/race/domain/RaceTest.kt
@@ -1,4 +1,4 @@
-package race
+package race.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
